Fixing comment counts to display properly
The previous fix fixed comment counts to be displayed only the in post's
page itself and not in page such as the home page of the blog etc (where
you only saw a "comments" like without the number of comments).

This commit fixes it by adding our very own javascript that fetches all
the comment counts when any page loads and updates the links
accordingly.

What more, because we now call the FQL server ourselves, we have a
callback to rely on in order to make sure that the correct pluralization
of the comments count is displayed (e.g. '1 comment' vs '2 comments')